+++ src/ignitus-Shared/ignitus-DesignSystem/ignitus-Layout/ignitus-NotFound/style.ts
@@ -3,7 +3,7 @@ import {keyframes} from '@emotion/core';
 import * as C from '../../ignitus-Atoms/colors';
 import {SM} from '../../ignitus-Atoms/fonts';
 import {Paragraph} from '../../ignitus-Atoms/atoms';
-import {PrimaryButton} from '../../ignitus-Atoms/ignitus-Buttons/Components/index'
+import {PrimaryButton} from '../../ignitus-Atoms/buttons';
 
 export const eye = keyframes`
   0% ,30% , 55%, 90% , 100% {
